requirements.txt
streamlit>=1.31.0
trueskill>=0.4.5
pandas>=2.0.0
PyGithub==1.58.2